[sint] Allow tests to be ordered
This commit introduces a new optional configuration property (`executionOrder`) that can be used to apply a strict order to the execution of tests.

Two options are provided:
- 'alphabetic'
- 'reversed'

Both sort the test based on the natural order of their declaring class name, followed by the test method name. The 'reversed' option used the reverse ordering of the 'alphabetic' option.

The new property is optional: when not defined, no explicit order is applied. Java's default implementation does not guarantee a specific order (and indeed, through experience we've learned that the order of test execution isn't consistent between test runs).

The rationale for this change is provided in https://github.com/XMPP-Interop-Testing/smack-sint-server-extensions/issues/114 : By being able to run tests in a specific order, and then once more in the reverse order, it becomes a lot easier to identify tests that have an undesired effect on each-other.